final refactor updates. ready for pull request
