+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import GameMode from "./GameMode";
+import GameMode from "./components/GameMode";
 
 function App() {
