+++ src/components/PresetDesign.js
@@ -2,8 +2,8 @@ import React, { useState, useReducer, useEffect, useContext } from "react";
 import * as Tone from "tone";
 import PlayPads from "./playpadMode/PlayPads";
 import "./PresetDesign.css";
-import TrackToggle from "../TrackToggle";
-import { initialState, isActiveReducer } from "./isActiveReducer";
+import TrackToggle from "./TrackToggle";
+import { initialState, isActiveReducer } from "../reducers/isActiveReducer";
 import { initialRecState, recordingsReducer } from "../reducers/recordingsReducer";
 import Sequencer from "./sequenceMode/Sequencer";
 import {gameModeData} from './GameMode';
